Rename folder SharedComponents to components
