+++ apps/judicial-system/web/src/routes/Prosecutor/Indictments/Processing/Processing.tsx
@@ -28,8 +28,8 @@ import CommentsInput from '@island.is/judicial-system-web/src/components/Comment
 import { isProcessingStepValidIndictments } from '@island.is/judicial-system-web/src/utils/validate'
 import * as constants from '@island.is/judicial-system/consts'
 
-import ProsecutorSection from '../../SharedComponents/ProsecutorSection/ProsecutorSection'
-import SelectCourt from '../../SharedComponents/SelectCourt/SelectCourt'
+import ProsecutorSection from '../../components/ProsecutorSection/ProsecutorSection'
+import SelectCourt from '../../components/SelectCourt/SelectCourt'
 
 const Processing: React.FC = () => {
   const {
