+++ apps/judicial-system/web/src/routes/Prosecutor/InvestigationCase/CaseFiles/CaseFiles.tsx
@@ -37,7 +37,7 @@ import * as constants from '@island.is/judicial-system/consts'
 import { removeTabsValidateAndSet } from '@island.is/judicial-system-web/src/utils/formHelper'
 import PoliceCaseFiles, {
   PoliceCaseFileCheck,
-} from '../../SharedComponents/PoliceCaseFiles/PoliceCaseFiles'
+} from '../../components/PoliceCaseFiles/PoliceCaseFiles'
 
 export const CaseFiles: React.FC = () => {
   const {
