+++ apps/judicial-system/web/src/routes/Prosecutor/InvestigationCase/Defendant/Defendant.tsx
@@ -41,8 +41,8 @@ import * as constants from '@island.is/judicial-system/consts'
 
 import PoliceCaseNumbers, {
   usePoliceCaseNumbers,
-} from '../../SharedComponents/PoliceCaseNumbers/PoliceCaseNumbers'
-import DefendantInfo from '../../SharedComponents/DefendantInfo/DefendantInfo'
+} from '../../components/PoliceCaseNumbers/PoliceCaseNumbers'
+import DefendantInfo from '../../components/DefendantInfo/DefendantInfo'
 
 const Defendant = () => {
   const router = useRouter()
