+++ apps/judicial-system/web/src/routes/Prosecutor/RestrictionCase/StepOne/StepOneForm.tsx
@@ -21,8 +21,8 @@ import * as constants from '@island.is/judicial-system/consts'
 
 import PoliceCaseNumbers, {
   usePoliceCaseNumbers,
-} from '../../SharedComponents/PoliceCaseNumbers/PoliceCaseNumbers'
-import DefendantInfo from '../../SharedComponents/DefendantInfo/DefendantInfo'
+} from '../../components/PoliceCaseNumbers/PoliceCaseNumbers'
+import DefendantInfo from '../../components/DefendantInfo/DefendantInfo'
 
 interface Props {
   workingCase: Case
