+++ apps/judicial-system/web/src/routes/Prosecutor/components/ProsecutorSection/ProsecutorSectionHeading.tsx
