+++ apps/judicial-system/web/src/routes/Prosecutor/components/ProsecutorSection/ProsecutorSectionHeightenedSecurity.strings.ts
