+++ apps/judicial-system/web/src/routes/Prosecutor/components/ProsecutorSection/ProsecutorSelection.strings.ts
